Refactor Reducer+Presents extension
