+++ Sources/ComposablePresentation/Reducer+PresentsKeyPath.swift
@@ -1,0 +1,31 @@
+import ComposableArchitecture
+
+extension Reducer {
+  /// Combines the reducer with another reducer that works on optionally presented `LocalState`.
+  ///
+  /// - All effects returned by another reducer will be canceled when `LocalState` becomes `nil`.
+  ///
+  /// - Parameters:
+  ///   - localReducer: A reducer that works on `LocalState`, `LocalAction`, `LocalEnvironment`.
+  ///   - toLocalState: A key path that can get/set `LocalState` inside `State`.
+  ///   - toLocalAction: A case path that can extract/embed `LocalAction` from `Action`.
+  ///   - toLocalEnvironment: A function that transforms `Environment` into `LocalEnvironment`.
+  /// - Returns: A single, combined reducer.
+  public func presents<LocalState, LocalAction, LocalEnvironment>(
+    _ localReducer: Reducer<LocalState, LocalAction, LocalEnvironment>,
+    state toLocalState: WritableKeyPath<State, LocalState?>,
+    action toLocalAction: CasePath<Action, LocalAction>,
+    environment toLocalEnvironment: @escaping (Environment) -> LocalEnvironment
+  ) -> Self {
+    combined(
+      with: localReducer.optional().pullback(
+        state: toLocalState,
+        action: toLocalAction,
+        environment: toLocalEnvironment
+      ),
+      cancelEffects: { state in
+        state[keyPath: toLocalState] == nil
+      }
+    )
+  }
+}
